finish user auth & setup ui for admin panel

+++ Angular/TunRW-app/backend/routes/events.js
@@ -117,9 +117,12 @@ router.put('/:id',checkAuth, multer({storage : storage}).single("image"), (req, 
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Event.deleteOne({_id:req.params.id}).then(result =>{
-    console.log(result);
-    res.status(200).json({  message : "Post deleted!"});
+  Event.deleteOne({_id:req.params.id, creator: req.userData.userId}).then(result =>{
+    if(result.n > 0){
+      res.status(200).json({ message: "Event Deleted !"});
+    }else {
+      res.status(401).json({  message : "Not Authorized!"});
+    }
   });
 
 });
